Quarantine D1 faction badge logic in a single file
