+++ src/app/inventory/store/d1-item-factory.ts
@@ -1,3 +1,4 @@
+import { factionItemAligns } from 'app/destiny1/d1-factions';
 import { t } from 'app/i18next-t';
 import {
   DestinyAmmunitionType,
@@ -30,31 +31,6 @@ const tiers = ['Unused 0', 'Unused 1', 'Common', 'Uncommon', 'Rare', 'Legendary'
 let _idTracker: { [id: string]: number } = {};
 // A map from instance id to the last time it was manually moved this session
 const _moveTouchTimestamps = new Map<string, number>();
-
-// Nodes that require matching faction alignment
-const factionNodes = {
-  652505621: 'New Monarchy',
-  2669659850: 'Future War Cult',
-  2794386410: 'Dead Orbit',
-};
-
-/**
- * Check to see if this item has a node that restricts it to a
- * certain faction, and if the character is aligned with that
- * faction.
- */
-function factionItemAligns(store: D1Store, item: D1Item) {
-  if (!item.talentGrid) {
-    return true;
-  }
-
-  const factionNode = item.talentGrid.nodes.find((n) => factionNodes[n.hash]);
-  if (!factionNode) {
-    return true;
-  }
-
-  return factionNodes[factionNode.hash] === store.factionAlignment();
-}
 
 // Prototype for Item objects - add methods to this to add them to all
 // items.
